requirements.txt
# requirements.txt
    
python-dotenv
chromadb
tiktoken
unstructured[pdf]            
unstructured-client        
sentence-transformers
transformers
accelerate
pillow
lxml
#  PyTorch  which is compatible with CUDA 12.1 onwards and also with python 3.9 and 3.10 not sure about other versions.
#make sure your pytorch packages is compatible with cuda version installed in your system which in turen is compatible with your python version.
torch==2.5.1+cu121
torchvision==0.20.1+cu121
torchaudio==2.5.1+cu121
--extra-index-url https://download.pytorch.org/whl/cu121
